requirements.txt
asgiref==3.9.0
attrs==25.3.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==42.0.8
defusedxml==0.7.1
Django==5.2.4
django-extensions==3.2.3
django-jazzmin==3.0.1
django-rql==4.4.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lark-parser==0.11.0
lib-rql==2.0.2
MarkupSafe==3.0.2
oauthlib==3.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
six==1.17.0
social-auth-app-django==5.4.0
social-auth-core==4.7.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.0.2
